Moved Score.java to Main Package
