+++ app/src/main/java/com/example/draw4brains/games/connectthedots/controller/DatabaseMgr.java
@@ -6,8 +6,7 @@ import android.util.Log;
 import androidx.annotation.NonNull;
 
 import com.example.draw4brains.games.connectthedots.model.Node;
-import com.example.draw4brains.games.connectthedots.model.Score;
-import com.example.draw4brains.games.connectthedots.view.GameMenuActivity;
+import com.example.draw4brains.main.model.Score;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
